[ GW ] tech debt repayment : moving context handling (delete) function to separete .h and .c from xively.c to make other internals use it (GW layerchain, and libxively driver for ftests already was using these function predeclared: maintenance critical)
