+++ src/tests/itests/xi_itest_helpers.h
@@ -149,20 +149,6 @@ extern xi_state_t xi_create_context_with_custom_layers( xi_context_t** context,
                                                         size_t layer_chain_size );
 
 /**
- * @brief xi_delete_context_with_custom_layers
- *
- * hidden API function that allows us to delete context with custom
- * layer setup
- *
- * @param context
- * @param layer_config
- * @return
- */
-extern xi_state_t xi_delete_context_with_custom_layers( xi_context_t** context,
-                                                        xi_layer_type_t layer_config[],
-                                                        size_t layer_chain_size );
-
-/**
  * macros for extracting next, and prev layer user data - this will be replaced
  * by the lookup function
  */
